% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bam_to_cov.R
\name{convert_bam_to_cov}
\alias{convert_bam_to_cov}
\title{convert_bam_to_cov}
\usage{
convert_bam_to_cov(
  bam_file,
  covs,
  map_profile = mappability_profile,
  paired = F
)
}
\arguments{
\item{bam_file}{path to BAM file(s)}

\item{covs}{a list of Coverage profile(s) in RLE format. Can be one or more samples.}

\item{map_profile}{The mappability profile from a host genome (we only have
Arabidopsis thaliana right now)}

\item{paired}{Indicate if the sequencing paired is single-end or paired-end
reads. True if paired-end. False if single-end.}
}
\value{
A data frame object that contains the mapping result for each virus
segment that the plant sample reads are aligned to and a RLE list of coverage
information
}
\description{
Converts one or more indexed and sorted BAM files into a
feature-extracted data frame after mappability profiling
}
