% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_high_nucleotide_content.R
\name{create_high_nucleotide_content}
\alias{create_high_nucleotide_content}
\title{create_high_nucleotide_content}
\usage{
create_high_nucleotide_content(gc = 0.6, a = 0.45, window = 75, virus_info)
}
\arguments{
\item{gc}{The threshold for GC content. It is the proportion of G and C
nucleotides in a sliding window. Default is 0.6.}

\item{a}{The threshold for A nucleotide. It is the proportion of A nucleotide
in a sliding window. Default is 0.45.}

\item{window}{The sliding window size of your choice. Default is 75.}

\item{virus_info}{A DNAStringSet of virus segments. The format should be similar to \code{virus_segments}.}
}
\value{
A data frame of the start and end positions of the regions_a that are
considered high in A\% and GC\%.
}
\description{
Creates a data frame of the start and end positions of the
regions_a that are considered high in A\% and GC\%.
}
\examples{
\dontrun{high_nucleotides_regions <- create_high_nucleotide_content()}

}
