% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cov}
\alias{example_cov}
\title{Coverage profiles of three plant samples.}
\format{
A list of 3 run length encoding (RLE) lists for 3 plant samples.
Each RLE list has the RLE vector of a virus segment
}
\usage{
example_cov
}
\description{
A list of coverage profiles for three plant samples. This is only a toy sample.
You can use it for running the examples in the vignette. We recommend using
more data to train the model, the more the better.
}
\keyword{datasets}
