% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_diag}
\alias{example_diag}
\title{Known diagnostics result of virus segments}
\format{
A matrix with 3 columns:
\describe{
\item{S1}{Sample one}
\item{S2}{Sample two}
\item{S3}{Sample three}
}
}
\usage{
example_diag
}
\description{
A matrix containing the known truth about the diagnostics result (using virus
database version 1.4.0) for each plant sample for the example data. It
records whether the sample is infected with a virus segment. Each column is a
sample, and each row is a virus segment's diagnostics status for three samples.
}
\keyword{datasets}
