% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mappability_profile}
\alias{mappability_profile}
\title{The mappability profile of the virus segments}
\format{
A list of 3,138 unmappable regions in the run-length encoding format.
}
\usage{
mappability_profile
}
\description{
A list of unmappable regions for the 3,138 virus segments.
}
\keyword{datasets}
