% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nucleotide_info}
\alias{nucleotide_info}
\title{Nucleotide information of 3,138 virus segments}
\format{
A data frame with 3,138 rows and 7 variables:
\describe{
\item{virus_name}{The virus name}
\item{iso_id}{The virus isolate ID}
\item{seg_id}{The virus segment ID}
\item{A_percent}{The percentage of A nucleotides in the virus segment}
\item{C_percent}{The percentage of C nucleotides in the virus segment}
\item{T_percent}{The percentage of T nucleotides in the virus segment}
\item{GC_percent}{The percentage of G and C nucleotides in the virus
segment}
\item{seg_len}{The length of the virus segment}
}
}
\usage{
nucleotide_info
}
\description{
A data set containing the GC content and other information about the virus
segments from the official Virtool virus data base. The variables are as
follows:
}
\keyword{datasets}
