% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cov.R
\name{plot_cov}
\alias{plot_cov}
\title{plot_cov()}
\usage{
plot_cov(
  covs,
  legend_status = TRUE,
  nucleotide_status = TRUE,
  window = 75,
  nucleotide_info_version = "1_4_0",
  ...
)
}
\arguments{
\item{covs}{An RLE list of coverage information of one or more plant samples.}

\item{legend_status}{Whether display legend. Default is \code{TRUE}.}

\item{nucleotide_status}{Whether display a sliding window of A percentage and
CG content. Default is \code{TRUE}.}

\item{window}{The sliding window size. Default is 75.}

\item{nucleotide_info_version}{The version number (character string) of the
nucleotide information of the virus segments. Default is \verb{1_4_0}.}

\item{\dots}{Other arguments that can be passed to \code{plot},
\code{lines}, or \code{legend}.}
}
\value{
The coverage profile of the mapped plant sample.
}
\description{
Plots the coverage profile of the mapped plant sample.
}
\examples{
plot_cov(example_cov$S1)



}
