% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_iimi.R
\name{train_iimi}
\alias{train_iimi}
\title{train_iimi()}
\usage{
train_iimi(
  train_x,
  train_y,
  method = "xgb",
  params_rf = list(ntree = 100, nodesize = 1, replace = T, mtry =
    floor(sqrt(ncol(train_x)))),
  nrounds_xgb = 100,
  params_xgb = list(max_depth = 10, gamma = 6)
)
}
\arguments{
\item{train_x}{A data frame or a matrix of predictors}

\item{train_y}{A response vector of labels (needs to be a factor)}

\item{method}{The machine learning method of choice, \verb{Random Forest} or
\code{XGBoost}, or \verb{Elastic Net}. Default is \code{XGBoost}.}

\item{params_rf}{A list of parameters to train a \verb{Random Forest} model in
\code{train_iimi()}}

\item{nrounds_xgb}{Max number of boosting iterations}

\item{params_xgb}{A list of parameters to train an \code{XGBoost} model in
\code{train_iimi()}}
}
\value{
A \verb{Random Forest}, \code{XGBoost}, \verb{Elastic Net} model
}
\description{
Trains a \code{XGBoost} (default), \verb{Random Forest}, or \verb{Elastic Net}
model using user-provided data. Default for \code{XGBooost} model is:
\itemize{
\item{\verb{nrounds = 100, max_depth = 10, gamma = 6}}
}
Default for \verb{Random Forest} model is:
\itemize{
\item{\verb{ntree = 100, nodesize = 1, replace = T, mtry = floor(sqrt(ncol(train_x))))}}
}
Default for \verb{Elastic Net} model is:
\itemize{
\item{\code{family = "binomial"}}
}
}
