% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trained_en}
\alias{trained_en}
\title{A trained model using the default Elastic Net settings}
\format{
An object of class \code{cv.glmnet} of length 13.
}
\usage{
trained_en
}
\description{
A trained model using the default Elastic Net settings
}
\keyword{datasets}
