% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trained_rf}
\alias{trained_rf}
\title{A trained model using the default Random Forest settings}
\format{
An object of class \code{randomForest.formula} (inherits from \code{randomForest}) of length 19.
}
\usage{
trained_rf
}
\description{
A trained model using the default Random Forest settings
}
\keyword{datasets}
