% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unreliable_regions}
\alias{unreliable_regions}
\title{The unreliable regions of the virus segments}
\format{
A data frame of unreliable regions in the run-length encoding format
for virus segments.
\describe{
\item{Start}{The start position of the region that is considered unreliable}
\item{End}{The end position of the region that is considered unreliable}
\item{Virus segment}{The virus segment ID}
\item{Categories}{The category that this unrelible region belong to, which
are Unmappable regions (host), Unmappable regions (virus), CG\% > 60\%, A\% > 45\%}
\item{1_4_0}{The version number of the virus database}
\item{1_5_0}{The version number of the virus database}
}
}
\usage{
unreliable_regions
}
\description{
A data frame of unmappable regions and regions of CG\% and A\% over 60\% and 45\%
respectively for the virus segments. It is worth to note that if a
virus segment does not have any unreliable regions, that virus segment is not
shown in this data frame.
}
\keyword{datasets}
