% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{virus_segments}
\alias{virus_segments}
\title{A DNAStringSet of virus segments from the Virtool virus data base (version 1.4.0)}
\format{
An object of class \code{DNAStringSet} of length 5531.
}
\usage{
virus_segments
}
\description{
A DNAStringSet of virus segments from the Virtool virus data base (version 1.4.0)
}
\keyword{datasets}
