% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMtocov.R
\name{BAMtocov}
\alias{BAMtocov}
\title{BAMtocov}
\usage{
BAMtocov(bamfile, covs, MP = mappability_profile, type)
}
\arguments{
\item{bamfile}{BAM file(s)}

\item{covs}{Coverage profile(s) in RLE format. If multiple samples, a
list of RLE mapping result for the samples.}

\item{MP}{The mappability profile from a host sample (we only have
Arabidopsis thaliana right now)}

\item{type}{Indicate if the sequencing type is single-end or pair-end reads}
}
\value{
A data frame object that contains the mapping result for each virus
segment that the plant sample is aligned to and a RLE list of coverage
information
}
\description{
Converts one or more indexed and sorted BAM files into a
feature-extracted data frame after mappability profiling
}
