% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rle_to_df.R
\name{convert_rle_to_df}
\alias{convert_rle_to_df}
\title{Convert run-length encodings (RLEs) to a data frame of features.}
\usage{
convert_rle_to_df(
  covs,
  unreliable_regions = NULL,
  additional_nucleotide_info = data.frame()
)
}
\arguments{
\item{covs}{A list of Coverage profile(s) in RLE format. Can be one or more
samples.}

\item{unreliable_regions}{A dataframe containing annotated unreliable regions of the
mapped reference. The information provided must be a 4-column data frame
that follows the format of \code{combined_unreliable_regions}, sans columns \verb{1_4_0}
and \verb{1_5_0}.}

\item{additional_nucleotide_info}{Additional nucleotide information for virus
segments that are not included in \code{nucleotide_info}. The information
provided must be a data frame that follows the format of
\code{nucleotide_info}. Default is an empty data frame.}
}
\value{
A data frame object that contains the mapping result for each virus
segment that the plant sample reads are aligned to and a RLE list of
coverage information.
}
\description{
Converts a list of run-length encodings (RLEs) into a data frame
with 16 features after mappability profiling and nucleotide filtering.
}
\examples{
\dontrun{
df <- convert_rle_to_df(example_cov)
}

}
