% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_diag}
\alias{example_diag}
\title{Known diagnostics result of over 3,000 virus segments}
\format{
A data frame with 3,138 rows and 21 columns:
\describe{
\item{row}{Each row is the name of a virus segment}
\item{column}{Each column is the name of a plant sample}
}
}
\usage{
example_diag
}
\description{
A data set containing the known truth about the diagnostics result for each
plant sample. It records whether the sample is infected with a virus segment.
}
\keyword{datasets}
