% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_iimi.R
\name{predict_iimi}
\alias{predict_iimi}
\title{predict_iimi()}
\usage{
predict_iimi(newdata, method, trained_model)
}
\arguments{
\item{newdata}{A matrix or data frame that contains the features extracted
from the coverage profile using \code{convert_bam_to_cov()}.}

\item{method}{The machine learning method of choice, \verb{Random Forest},
\code{XGBoost}, or \verb{Elastic Net}. Default is \code{XGBoost} model.}

\item{trained_model}{The trained model. If not provided, default model is
used.}
}
\value{
A data frame of diagnostics result for each sample
}
\description{
Uses a machine learning model to predict the infection status
for the plant sample(s). User can use their own model if needed.
}
