% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_iimi.R
\name{predict_iimi}
\alias{predict_iimi}
\title{predict_iimi()}
\usage{
predict_iimi(newdata, method, trained_model, seg_level = F)
}
\arguments{
\item{newdata}{A matrix or data frame that contains the diagnostics result}

\item{method}{The machine learning method of choice, either Random Forest or
XGBoost. Default is the XGBoost model}

\item{trained_model}{The trained model. If not provided, default model is used}

\item{seg_level}{Whether to report the diagnostics result in a virus segment
level. If TRUE, the result data frame returns detection result for each
segment in each sample. Default is FALSE.}
}
\value{
A data frame of diagnostics result for each sample
}
\description{
Uses a machine learning model to predict the diagnostics result
for the plant sample(s). User can use their own model if needed.
}
