% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_iimi.R
\name{train_iimi}
\alias{train_iimi}
\title{train_iimi()}
\usage{
train_iimi(
  train_x,
  train_y,
  method = "xgb",
  params_rf = list(ntree = 100, nodesize = 1, replace = T, mtry =
    floor(sqrt(ncol(train_x)))),
  nrounds_xgb = 100,
  params_xgb = list(max_depth = 10, gamma = 6)
)
}
\arguments{
\item{train_x}{A data frame or a matrix of predictors}

\item{train_y}{A response vector of labels (needs to be a factor)}

\item{method}{The machine learning method of choice, either Random Forest or
XGBoost. Default is XGBoost.}

\item{params_rf}{A list of parameters to train a Random Forest model in train_iimi()}

\item{nrounds_xgb}{Max number of boosting iterations}

\item{params_xgb}{A list of parameters to train an XGBoost model in train_iimi()}
}
\value{
A Random Forest or a XGBoost model
}
\description{
Trains a XGBoost (default) or Random Forest model using
user-provided data. Default for XGBooost model is:
\itemize{
\item{nrounds = 100, max_depth = 10, gamma = 6}
}
Default for Random Forest model is:
\itemize{
\item{ntree = 100, nodesize = 1, replace = T, mtry = floor(sqrt(ncol(train_x))))}
}
}
