% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trained_xgb}
\alias{trained_xgb}
\title{A trained model using the default XGBoost settings}
\format{
An object of class \code{raw} of length 114445.
}
\usage{
trained_xgb
}
\description{
A trained model using the default XGBoost settings
}
\keyword{datasets}
