% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{virus_segments}
\alias{virus_segments}
\title{The virus segments from the official Virtool virus data base}
\format{
An object of class \code{DNAStringSet} of length 3138.
}
\usage{
virus_segments
}
\description{
A DNAStringSet object of the 3,138 virus segments.
}
\keyword{datasets}
